Cumulative before v0.6.1

+++ salesforce/tests/test_utils.py
@@ -2,7 +2,6 @@
 Tests for `salesforce.utils`
 """
 from django.test import TestCase
-from django.db import connections
 import unittest
 
 
@@ -22,27 +21,37 @@ class UtilitiesTest(TestCase):
     def test_lead_conversion(self):
         """
         Create a Lead object within Salesforce and try to
-        convert it, then clean all the generated objects.
+		convert it, convert/merge it with the information from a duplicit Lead,
+        then clean all the generated objects.
         """
-        lead = Lead(FirstName="Foo", LastName="Bar", Company="django-salesforce")
+        lead = Lead(FirstName="Foo", LastName="Bar", Company="django-salesforce",
+                    Street='Test Avenue 45')
         lead.save()
-
-        # Get a status name setting for converted Leads
-        cur = connections['salesforce'].cursor()
-        cur.execute("SELECT MasterLabel FROM LeadStatus WHERE IsConverted=true")
-        converted_status = cur.fetchone()['MasterLabel']
-
-        ret = convert_lead(lead, converted_status=converted_status)
-        print("Response from convertLead: " +
-                ', '.join('%s: %s' % (k, v) for k, v in sorted(ret.items())))
-        expected_names = set(('accountId', 'contactId', 'leadId', 'opportunityId', 'success'))
-        self.assertEqual(set(ret), expected_names)
-        self.assertEqual(ret['success'], 'true')
-
-        # Cleaning up...
-        # Deleting the Account object will also delete the related Contact
-        # and Opportunity objects.
-        account = Account.objects.get(pk=ret['accountId'])
-        account.delete()
-
-        lead.delete()   # FYI, ret['leadId'] == lead.pk
+        lead2 = Lead(FirstName="Foo", LastName="Bar", Company="django-salesforce",
+                     Phone='123456789')
+        lead2.save()
+        ret = None
+        try:
+            # convert the first Lead
+            ret = convert_lead(lead, doNotCreateOpportunity=True)
+            #print("Response from convertLead: " +
+            #        ', '.join('%s: %s' % (k, v) for k, v in sorted(ret.items())))
+            expected_names = set(('accountId', 'contactId', 'leadId', 'opportunityId', 'success'))
+            self.assertEqual(set(ret), expected_names)
+            self.assertEqual(ret['success'], 'true')
+            # merge the new Account with the second Lead
+            ret2 = convert_lead(lead2, doNotCreateOpportunity=True, accountId=ret['accountId'])
+            account = Account.objects.get(pk=ret['accountId'])
+            # verify that account is merged
+            self.assertEqual(ret2['accountId'], account.Id)
+            self.assertEqual(account.BillingStreet, 'Test Avenue 45')
+            self.assertEqual(account.Phone, '123456789')
+        finally:
+            # Cleaning up...
+            if ret:
+                # Deleting the Account object will also delete the related Contact
+                # and Opportunity objects.
+                account = Account.objects.get(pk=ret['accountId'])
+                account.delete()
+            lead.delete()   # FYI, ret['leadId'] == lead.pk
+            lead2.delete()
